Separate gene and TF searches into different UI tabs
Since gene-based and TF-based searching and results display should behave differently,
separate these into different UI tabs. Thus, remove the radio button that allows
selection of either gene-based or TF-based searching and write duplicate code (and
modify variable names, etc appropriately) to generate and display gene-based and
TF-based search results. In the gene-based search results, do not show the gene name
column in the results (since it is the same for each row). In the TF-based search
results, do not show the TF name column (for similar reason).

TO-DO:
- Remove R2 column from gene-based results and include R2 value elsewhere
- For gene-based search, show all results at once rather than paginated results
- In both tabs, give brief instructions (near drop-list on left-hand side)
  indicating that one can type a gene/TF or scroll and select gene/TF
- In both tabs, give brief instructions (near search field on right-hand side)
  indicating that a TF/gene can be typed to further filter results in the table